+++ coeus-python/src/nn/mod.rs
@@ -4,6 +4,7 @@ pub mod dropout;
 pub mod embedding;
 pub mod feedforward;
 pub mod linear;
+pub mod module_base;
 pub mod module_list;
 pub mod normalization;
 pub mod pool;
@@ -18,6 +19,7 @@ pub use dropout::PyDropout;
 pub use embedding::PyEmbedding;
 pub use feedforward::{PyFeedForward, PyTransformerDecoderLayer};
 pub use linear::PyLinear;
+pub use module_base::PyModule;
 pub use module_list::PyModuleList;
 pub use normalization::{
     PyBatchNorm1d, PyBatchNorm2d, PyBatchNorm3d, PyGroupNorm, PyInstanceNorm1d, PyInstanceNorm2d,